Refactor BudgetPage
