+++ client/src/ui/atoms/ActionHeader.jsx
@@ -9,7 +9,6 @@ export const ActionHeader = ({
 }) => {
   return (
     <Box
-      // paddingBottom={3}
       spacing={{
         xs: 3,
         md: 0,
